Refactor save/load preferences values functions
