+++ newIDE/app/src/MainFrame/Preferences/PreferencesContext.js
@@ -2,6 +2,8 @@
 import { Trans } from '@lingui/macro';
 import * as React from 'react';
 
+const LocalStorageItem = 'gd-preferences';
+
 export type AlertMessageIdentifier =
   | 'use-non-smoothed-textures'
   | 'use-nearest-scale-mode'
@@ -131,6 +133,8 @@ export type Preferences = {|
   setUseGDJSDevelopmentWatcher: (enabled: boolean) => void,
   setEventsSheetUseAssignmentOperators: (enabled: boolean) => void,
   setShowEffectParameterNames: (enabled: boolean) => void,
+  loadPreferencesValues: () => PreferencesValues,
+  savePreferencesValues: (preferences: Preferences) => Preferences,
 |};
 
 export const initialPreferences = {
@@ -164,7 +168,44 @@ export const initialPreferences = {
   setUseGDJSDevelopmentWatcher: (enabled: boolean) => {},
   setEventsSheetUseAssignmentOperators: (enabled: boolean) => {},
   setShowEffectParameterNames: (enabled: boolean) => {},
+  loadPreferencesValues: loadPreferencesValues,
+  savePreferencesValues: savePreferencesValues,
 };
+
+export function loadPreferencesValues(): ?PreferencesValues {
+  try {
+    const persistedState = localStorage.getItem(LocalStorageItem);
+    if (!persistedState) return null;
+
+    const values = JSON.parse(persistedState);
+
+    // "Migrate" non existing properties to their default values
+    // (useful when upgrading the preferences to a new version where
+    // a new preference was added).
+    for (const key in initialPreferences.values) {
+      if (
+        initialPreferences.values.hasOwnProperty(key) &&
+        typeof values[key] === 'undefined'
+      ) {
+        values[key] = initialPreferences.values[key];
+      }
+    }
+
+    return values;
+  } catch (e) {
+    return null;
+  }
+}
+
+export function savePreferencesValues(values: PreferencesValues) {
+  try {
+    localStorage.setItem(LocalStorageItem, JSON.stringify(values));
+  } catch (e) {
+    console.warn('Unable to persist preferences', e);
+  }
+
+  return values;
+}
 
 const PreferencesContext = React.createContext<Preferences>(initialPreferences);
